Refactor protocols package into modular structure
