+++ protocols/profiles/dream_weaver.py
@@ -1,0 +1,8 @@
+"""Creative patch-suggestion agent profile."""
+from protocols.core.profiles import AgentProfile
+
+DreamWeaver = AgentProfile(
+    name="DreamWeaver",
+    traits=["creative", "collaborative"],
+    powers=["propose_patch", "ideate"],
+)
